+++ packages/uploadthing/test/__test-helpers.ts
@@ -307,50 +307,3 @@ export const it = itBase.extend<{ db: MockDbInterface }>({
     files.length = 0; // clear files after each test
   },
 });
-
-/**
- * Starts an Express server and sets up mocked handlers using MSW
- */
-export const setupUTServer = () => {
-  const f = createUploadthing({
-    errorFormatter(err) {
-      // eslint-disable-next-line no-console
-      console.log(err, err.cause);
-      return { message: err.message };
-    },
-  });
-  const router = {
-    foo: f({ text: {} })
-      .middleware((opts) => {
-        middlewareMock(opts);
-        return {};
-      })
-      .onUploadError(onErrorMock)
-      .onUploadComplete(uploadCompleteMock),
-  };
-
-  const app = express();
-  app.use(
-    "/api/uploadthing",
-    createRouteHandler({
-      router,
-      config: {
-        uploadthingSecret: "sk_test_123",
-        isDev: true,
-      },
-    }),
-  );
-
-  const server = app.listen();
-  const port = (server.address() as { port: number }).port;
-
-  const uploadFiles = genUploader<typeof router>({
-    package: "vitest",
-    url: `http://localhost:${port}`,
-  });
-
-  return {
-    uploadFiles,
-    close: () => server.close(),
-  };
-};
